[V11.2.1] In RecurrentEdgeProcess class, generate hourly quantities over number of parallel edge usage journeys from edge usage pattern instead of generating over number of hourly edge usage journey starts.
In fact, the number of parallel edge usage journeys takes into account the usage span of the edge usage journey, while the number of hourly edge usage journey starts doesn’t. This bug fix leads to significant changes in the results when the edge usage journey duration is long compared to the length of edge usage journey starts (which can often be the case).